v 0.7
finally finished with xsd and xml export and import. it took a lot of time.

There is appearing a bug after dropping all objects from database, if you create new objects for models and then make export and import - everything will be fine, but for some reasona after short period of time, if you would check once is it working or not by deleating already existing player-1.xml and .json import log, after running the program you will face with the issue with status 500 for both methods.

I don't know what happens, but it can be fixed only by refreshing the database