v 0.8
started doing UNIT-tests. for now  there is only those for modules, that can only check for correct entering data in fields, but everything starts from small things.

For today i am very tired already, so i will continue tomorrow in the day.

To finish this project there is only remained to write other UNIT-tests, make JavaDoc where it is needed and may be for real polishing i need to finish controller requests, so all getters and setters would be painted blue, not grey. But, still, may be it is not that important, due to all the basic stuff is implemented and those methods are just for fun, just to write them and see how they are working. I will find out that later!